#82 Написал скрипт добавления новых тестовых данных в БД

+++ db_migrations/changesets/rollback/M-010.rollback.sql
@@ -1,5 +1,5 @@
-DELETE FROM matcher.project_participation WHERE project_id >= 2147483643;
-DELETE FROM matcher.projects WHERE project_id >= 2147483643;
-DELETE FROM matcher.user_organizations WHERE org_id >= 2147483644;
-DELETE FROM matcher.organizations WHERE org_id >= 2147483644;
-DELETE FROM matcher.users WHERE user_id >= 2147483638;
+DELETE FROM matcher.project_participation WHERE project_id >= 2147483643 AND project_id <= 2147483647;
+DELETE FROM matcher.projects WHERE project_id >= 2147483643 AND project_id <= 2147483647;
+DELETE FROM matcher.user_organizations WHERE org_id >= 2147483644 AND org_id <= 2147483647;
+DELETE FROM matcher.organizations WHERE org_id >= 2147483644 AND org_id <= 2147483647;
+DELETE FROM matcher.users WHERE user_id >= 2147483638 AND user_id <= 2147483647;
